Add looping job titles
